(maint) Fix failing integration metrics test
Adds FIXME comments to describe some undesirable behavior we are seeing
the metrics because they are JVM global and persisting between tests.

Something between the commits `118756370f` and `4c0b8e038f` has caused
the tests in puppetlabs.puppetdb.integration.masterless to start calling
`["puppetdb" "facts" "find"]` (only on the `main` branch? this hasn't
failed on 6.x). It is likely caused by changes to how the environment is
loaded in PUP-11328, but it's unclear why it is not failing when testing
PuppetDB against puppetlabs/puppet `6.x`.

```
*   118756370f Josh Cooper 3 weeks ago Merge pull request #8803 from joshcooper/6x_main_merge_oct26
|\
| * e5fcef44cb Josh Cooper 3 weeks ago Merge remote-tracking branch 'upstream/6.x' into 6x_main_merge_oct26
|/|
| *   bb1017f923 Josh Cooper 3 weeks ago Merge pull request #8802 from GabrielNagy/PUP-11328/node-request-fallback
| |\
| | * 867691b617 Josh Cooper 3 weeks ago (PUP-11328) Get the environment's name without loading the environment
| | * 9d3e630a67 Gabriel Nagy 3 weeks ago (PUP-11328) Fallback to node request in configurer
| |/
| * aaae1177c9 Jenkins CI 3 weeks ago (packaging) Updating the puppet.pot file
| *   88905a5505 Gabriel Nagy 3 weeks ago Merge pull request #8784 from luchihoratiu/PUP-11204
| |\
| | * 430ffa30cc Luchian Nemes 4 weeks ago (PUP-11204) Allow `puppet lookup --facts` to receive any filename  (HEAD -> main)
* | | 4c0b8e038f Jenkins CI 3 weeks ago (maint) Merge up 45def53 to main
```