(PDB-5378) Get latest report timestamp from certnames table
Scanning the report partitions for the latest report timestamp is
terribly inefficient. We have the latest report timestamp in the
certnames table so just use that.

This check needs to filter out the row when latest_report_timestamp is
NULL because NULL sorts (desc) before real values. So this check would
always return null for a certname with no reports. Instead, omit the row
and let the query decide between the factset/catalog timestamps if
present.

```sql
select * from (VALUES(NULL), (NOW())) AS foo(timestamp) order by timestamp desc limit 1;
```