(PDB-5282) Fix handling of empty certname at command endpoint
A command with an empty string for a certname would cause the command
endpoint to throw unmanaged exceptions and potentially lock up syncing
PuppetDB applications thereafter. This was due to the way metadata was
being stored and read back in the queue. A metadata filename would be
happily created with no certname but upon processing would not recognize
the metadata without the certname. In particular, the regex did not
match empty certnames, so the result of the regex search and metadata
parsing would return nil.

In order to prevent this from happening, this change updates the command
http endpoint by correcting issues with request validation. Previously
some validation was done with Schema definitions, but the error results
returned were very unfriendly to our users. Some exceptions were not
caught at all until Jetty caught them resulting in status 500 HTML
responses for our client who just sent us JSON. Other Schema validation
errors returned huge blobs of indecipherable error messages in the
response JSON. This patch removes all Schema validation from the command
endpoint, creates a new middleware that validates the incoming requests
with human friendly error messages, and adds an exception handling
middleware to the "front" of our command endpoint middleware stack.

This patch also refactors the request normalization middleware into a
single cohesive function instead of the three functions that were
previously responsible for shaping the "old" style and "new" style into
the same structure. The result is a much clearer picture of how requests
get normalized.

The exception handling issue in PDB-5282 can be described as such:
command/process-message receives nil for the cmdref argument due to the
aforementioned metadata parsing failure. Destructuring occurs anyways
resulting in nil command, version, certname, etc. When command and
version are nil that function throws an exception when it tries to
update the metrics because metrics needs the command and version name.
This problem is not addressed in this patch as it is an architectural
weakness of the current queue/metadata/metrics design. Needs further
attention.