(PDB-5381) Ensure that stream-query-result passes the proper query config
Sync is bouncing nodes, and creating unnecessary calls to the previously
expensive `has-newer-record-for-certname?`. Because it has the record
already, the check is true and it does not enqueue the command. This
results in a count of 0 for the nodes entity during sync even though it
did all the work of transferring and vetting a command, but just didn't
enqueue it.

The discrepancy between the local and remote summary queries is between
the handling of ttls between produce-streaming-body which gets called by
the HTTP(S) query endpoint and stream-query-result which gets called by
the trapperkeeper endpoint.

produce-streaming-body properly handles the context, and turns it into
query-config as expected

```
query-config (select-keys context [:node-purge-ttl :add-agent-report-filter])
```

In steam-query-result the node-purge-ttl setting needed to properly set
the inactive_nodes CTE for the query is in the context. But we strip it
out before proceeding to turn the AST into SQL.

```
(select-keys context [:log-queries :query-id])
```

Passing the proper config values through stream-query-result broke the
export of plan reports because it was relying on add-agent-report-filter
being absent. By properly passing that config value, a `["from"
"reports"]` query gets a default filter of `["=" "type" "agent"]` added
and omits non-agent reports. This modifies the export query to
explicitely query for all reports.