(PDB-5333) Allow retrieving PDB version without a db query
Previously, our status endpoint always included database information.
However, generating these database statuses is expensive and we want to
provide a way to get the PDB version without hitting the database.

The Trapperkeeper service trapperkeeper-status is used to provide the
status endpoint at /status/v1/services/puppetdb-status. The service
supports multiple forms of status responses via the "level" query
parameter. Before this change we ignored the level. Now we check the
level and we do not provide database statuses if it is :critical. The
levels :info (default) and :debug return the full status information
including the databases statuses.

The status endpoint responses at the default level :info are exactly the
same so this is not a breaking change.

This commit also removes runtime validation for the create-status-map
function because it is of questionable usefulness. Type errors will
surely be caught during testing because this function does not interact
with user input directly.