(PDB-4285) Improve command endpoint docs
This fixes docs about the command endpoint which were confusing
regarding the "new" and the "old" command format. The docs have been
changed to explicitly require that a certname be included with every
command, although this requirement may change in the future as we add
commands that do not correspond to only a single certname.