Revert "(PDB-5333) Allow retrieving PDB version without a db query"
This reverts commit f1c536f32a2f4814aee4d92cf0b4d9aa371bca25.

This commit was not tested against PE PDB and causes failures in PE PDB.
PE PDB relies on the implicitly exposed API of PDB functions involved
with creating the status endpoint and refactoring PE PDB to use the
simplified interface that the reverted commit creates would be too
costly. The reverted commit should be reworked such that no API
breakages occur.