(PDB-5452) Disable postgres jit for all connections
Problems with the jit have been reported with the jit, which was added
in pg 11, and enabled by deafult in 12.  In particular, it can cause
some queries to run *much* more slowly, even in situations where there
don't appear to be many rows involved.

So until the problem is resolved, disable the jit for all pdb
connections, by adding a "set session jit = off;" to the hikaricp
connection init string.