(PT-1027) Add support for "strictly ordered" migrations
Rework migration handling to allow some migrations to be marked as
having to be run an a :strict-order?, in which case pdb will refuse to
start if the newest applied migration is newer than the newest pending
strict migration.  See the changes to the migrate namespace
documentation for some additional details.

This will be used for migrations that support experimental features,
so that if you've always had a given feature set in
PDB_EXPERIMENTAL_FEATURES, then the migrations related to that feature
will be applied correctly, whenever needed, but if you accidentally
enable a feature "after the fact", i.e. after the database has moved
past any of the feature's migrations, you'll be protected from
manipulating the database when it's in an unexpected, and potentially
incompatible state.