(PDB-5260) Avoid recursive rules to mitigate slow PQL parsing
Avoid recursion in some of our rules to avoid *very* bad performance
for, as an example, a large number of 'or' filter clauses.

Before the change, this PQL query would saturate the CPU for many
minutes before running out of memory and crashing the test program:

  (str "nodes[] { certname in [\""
       (str/join "\", \""
                 (repeat 5000 "heavy-empire.delivery.puppetlabs.net"))
       "\"] }"))

After the change, that same query finishes in a couple of seconds on a
fairly slow laptop.  While that still seems slow, it's a huge
improvement.

For what it's worth, for that particular query, the important change
is the one to <literallist>.

From https://github.com/Engelberg/instaparse/blob/4d1903b059e77dc0049dfbc75af9dce995756148/docs/Performance.md:

  Prefer using * and + over recursion to describe simple
  repetition. For example, the rule:

   <A> = 'a'+

  can be internally optimized in ways that

   <A> = 'a' A | 'a'

  cannot.