(PDB-5484) Test GET *and* POST in places it was overlooked
Unify the multiple versions of is-query-result and adjust it to avoid
evaluating the expected value twice, to use query-response, and to
pass it the method.  It appears this was overlooked when POST was
originally added 6b93f9b2639be1863197cb7fe63b82604c960bfe.

Use vector-params in a number of cases where it's needed now that
we're testing POST.