(HDP-161) Remove `submit_only` server URL when it times out repeatedly
This commit adds logic to PDB's Ruby http utils to track timeouts for
the `submit_only` server URLs, which are used by HDP. If a given URL
times out twice in a row, it is removed from the loaded config, to avoid
tying up JRubies indefinitely in the case where HDP has gone down. By
removing the URL, PDB will no longer try to submit things to it,
allowing requests to the other configured URLs to proceed quickly.

The URL will be re-added to the config whenever the JRuby instance is
refreshed: with a restart of Puppet Server, when the JRuby instance is
recreated when the "max requests per instance" threshold is hit, or if
the pool is flushed manually. Otherwise, the misbehaving URL will be
kept out of the list permanently.