Revert "(PDB-5333) Allow retrieving PDB version without a db query (v2)"
This reverts commit 27b68589aa3489e8d1bc301721a5297cd02b188c.

This commit is being reverted because it unexpectedly introduced a
breaking change. Before the reverted commit, the status endpoint did not
differentiate between levels. This means that clients could issue a
request to the status endpoint with any level ("debug", "info", or
"critical") and get the same response back. This singular response was
implicitly the "info" level, but some clients were requesting a
"critical" level status and expecting an "info" level response back.
Returning less information in a status response of any level will be
a breaking change because clients may be expecting that missing
information to be there.

The PDB docs don't mention the levels directly, they state that the
status endpoint implements the Puppet Enterprise Status API
(https://puppet.com/docs/pe/latest/status_api.html).