(PE-33893) Improve CONTRIBUTING.md document
Adds more information about the PostgreSQL sandbox inside of the
PuppetDB sandbox. Informs users that they can use pgbox directly.
Previously the PuppetDB persistent sandbox example was located inside
the repo and named test-sandbox. Now the example location is outside of
the repo and named pdb-sandbox for clarity about which kind of sandbox
it is.