(PE-33893) Add ci/README.md overview of testing and CI scripts
Previously there was no overview of the CI and testing scripts which are
mostly written in Bash. It can be hard to remember what they do. This
document adds brief descriptions for most CI and testing scripts and
should be updated when any major functionality of the scripts change and
when new scripts are added.