(PDB-5519) Convert honeysql v1 syntax to v2
Updates many honeysql v1 function calls to their new v2 format. For
example, `(hcore/raw str...)` becomes `[:raw str...]`, `hcore/inline`
becomes `[:inline ...]`, etc. In a few places we need to get the string
out of a SqlRaw type (so we use `:s`), now this is the second field in
an array.

Enable native support for postgresql operators by including
`honey.sql.pg-ops`. Where trivial (and supported by our current
parameter handling), some raws are converted to native honeysql v2
syntax now that their operators are supported natively. Most are left
raw, but honeysql should support all our use-cases natively.  Checking
for a raw value is different due to the format change as well.

Update the type coercion matrix. Turning a column into a raw statement
(even the new `[:raw _]` form) is not the "noop" it was being used for
in the type coercion matrix. It causes issues when the coerced type is
used in a selection field because `{:select [^SqlRaw "string"]}` would
become `{:select [[:raw "string"]]}` which compiles as "SELECT raw as
\"string\"" instead of the desired outcome. Instead we use keyword to
"convert" the column into itself. It should be noted that this is _only_
safe due to the fact that we only use type coercion on the column name.

Update `sql-cast` which instead of creating raw SQL calls for the
conversions, uses the built in method of calling functions in honeysql
v2.

`sql-regexp-array-match-v2` is added as the new honeysql v2 version to
avoid making changes to the legacy query engine.