(PDB-5519) Move paging handling to honeysql
Before, we were compiling paging options manually and tacking them onto
the end of a query. Limit and offset will always be an integer, but
order by could include honeysql to represent the field that we are
grouping by. Honeysql v2 doesn't compile partial expressions, and
instead errors. Therefore we instead pass the paging options through
with our generated query plan and allow honeysql to create the paging
options in the query directly.