(PDB-5524) Log command producer timestamp, hash, and outcome
Augment the "command processed for CERTNAME" messages to make it
easier to diagnose related issues.

Adjust all of the command "execution" functions to return a status map
containing the relevant information.

Change update-facts! to include the factset :hash unconditionally
since the code that needed it to be optional no longer exists.