(maint) dep-diff: allow ignoring bouncycastle provided scope
In Puppet Enterprise, the bouncycastle jars are not packaged in the
uberjar and are instead provided on the classpath depending on whether
or not it is a FIPS environment.

This allows the dep-diff script to ignore the provided scope on the
bouncycastle jars so that it just ensures we compare the versions of
bouncy castle.