cli/benchmark: improve startup time
In 75fb04f41 ("(maint) Ensure the benchmark tool creates the requested
number of hosts"), the benchmark startup was intentionally paused until
the full queue of nodes was populated to prevent the simulator from
non-uniformly distributing requests over the requested number of nodes.

For large number of nodes, say more than 10,000, this creates an
undesirable delay in the startup time of the overall benchmark.

This change uses a mix to ensure the initial hosts are read before
attempting to use the update channel in a loop. This allows immediate
startup but retains the uniform distribution we want.