cli/benchmark: improve TempFileBuffer performance
The operations in TempFileBuffer are performed under a single mutex. In
this case, it meant that channel put operations were essentially bounded
by the time taken to write the data file.

By moving the write to a separate asynchronous task, we're able to
achieve a ~4x increase in benchmark client throughput.