cli/benchmark: improve rate monitor reporting
The rate monitor reports an "equivalent load" as the benchmark runs, but
it isn't clear what that equivalent load is relative to. In particular,
in the case of --nummsgs, it uses a 30 minute comparative default. But
for --runinterval, it uses the actual run interval specified, which
isn't really helpful.

Most customers use a 30 minute run interval, so this change makes the
rate monitor always report in those terms, and makes it clear from the
output message that it is doing that.