(PDB-5555) Lock tables consistently when inserting and purging reports
Previously, we acquired table locks in reverse order when inserting vs.
purging reports. Insertion was certnames, reports, resource_events and
purging was resource_events, reports, certnames. This created the
possibility of a deadlock when two concurrent transactions each acquired
a portion of their locks at the same time. We now acquire all the locks
for report purging up front.

This introduces a helper function `acquire-locks!` which will lock a
given set of tables with specified modes, in a consistent order across
all callers. We now have a partial ordering (so far only consisting of
certnames, reports and resource_events) for tables, so that explicit
locks will always be acquired the same way.