(PDB-5559) Use both read and write db in all tests
Previously, many tests specified a single database which would be used
as both the read and write databases, but to fix a problem with
partition dropping (specifically the gc bulldozer) we need to add the
read user role to the write user.  But then, whenever a read
database (and user) isn't specified in the test config, the migration
coordination checks at startup will fail (crashing pdb) because the
coordinator will only revoke the connection rights from the write
user, and the write user will still have connect access indirectly via
the read role.

To fix that, specify both databases in all the test configs, which is
more appropriate anyway, since that's the recommended configuration.

As part of the work, revise call-with-test-dbs to pass both databases
to the function.