(PDB-5554) Detach partition concurrently
If we are working with a Postgresql 14+ database, we can use `detach
partition concurrently` which has narrower locking constraints compared
to dropping the table. This cannot run within a transaction, however, so
do-commands-outside-txn() is used and the start of the transaction for
the reports and resource-events gc is moved from services into the
respective delete-(reports|resource-events)-older-than! functions in
storage.clj.