(PDB-5559) Given a separate db read role; eject it during migrations
When the read database user isn't the same as the write database user,
eject it too during migrations.  Rework update-schema to handle/guard each
ejection/restoration independently, and rearrange the code to check
the state only after all of the changes.  Otherwise the checks will
fail if one of the users has connect access via another's role, which
is the case for the write user, which has the read user's role.