(PDB-5567) Add declarative partitioning migration for reports/resource_events
In order to be able to use ALTER TABLE DETACH PARTITION CONCURRENTLY, we
need to be using declarative partitioning for reports and
resource_events. This commit adds migrations transitioning the
inheritance partitioning for reports and resource_events to declarative
partitioning using the ALTER TABLE ATTACH PARTITION syntax.

We still need the original partitions generated in migrations #73/74, so
the functions for generating the inheritance partitions are moved into
migrate.clj. A new, somewhat simpler create-partition uses CREATE TABLE
PARTITION FOR VALUES FROM.

Indices are automatically generated on child partitions based on the
parent table's indices, so code generating indices on the children
manually is dropped.