(PDB-5568) Fix order by clause compilation
Prior to the upgrade to honeysql v2 (PDB-5519), we handled constructing
the order by clauses separately from the rest of the query. This meant
we needed to extract the sql from raw requests to create an SQL string
ahead of time. Now we are letting the paging components stay in the main
query until honeysql constructs the whole query at once at the end. This
means that during dealiasing we can simply select the honeysql field
from the query map with no futher alterations.