(PDB-5554) Separate PG11/14 workflow in storage partition delete
Prior to this both PG11 and PG14 databases would go through the same
partition delete workflow, but on PG14, partitions would be detached
before deletion. The problem was that both detaching and dropping the
tables went through prune-daily-partitions, which incrementally bases
what to detach or drop based on current get-partition-names. But that
method is querying children of the reports or resource_events table, and
since we've just detached one, the later call to drop doesn't see it and
drops the next table entirely. So detach and drop were out of sync.

To fix that I separated the PG11 workflow from the PG14, return
explicitly what we detached, and drop that list rather than relying on
what the state of some later query is.

There could be other issues here between an overlapping
periodic-gc and a manual sweep-reports! where their detaches could be
interleaved. Previously I don't think that was an issue because each
transaction would first have to acquire locks, but the new detach is
outside of the transaction. But Austin pointed out that periodic-gc and
a manual gc is thread locked in the clean-up function, so this shouldn't
be a problem.

Also made a change to the correctly-sweep-reports test. It was passing
{:incremental true} instead of {:incremental? true} and was ending up
with incremental null. The test actually needs to be run not incremental
to clear out enough days, so I switched this to an {incremental? false}.