(maint) Fix race in accessexclusivelocks test
Adds a CyclicBarrier before garbage collection starts to prevent the
start of the *next* gc until we are ready. Without this, when the test
awaits after-gc, it releases the invoke-periodic-gc to perform the next
gc cycle (which is every hundredth of a second by the test config) and
you get the race between the test validating which tables have dropped
and the gc dropping the next table.

Also adds a barrier right before query submission, and a small sleep to
hopefully let those threads submit their three queries. Otherwise,
timing can be such that the bulldozer can go by first, find nothing,
then the queries get submitted, then we hang for 20 min.

Unfortunately, we can't wait on a barrier absolutely after the query
submissions because the whole point is for those queries to hang and be
cleared by the bulldozer. We can't wait for them to come back.