(PDB-5613) Fix command serialization logging
Remove unused code from char_encoding, such as Utf8CharLens

Change error_char_context to not require a range argument

Remove utf8_string in favor of calling coerce_to_utf8 directly. The
utf8_string method called coerce_to_utf8 and caught a few encoding
errors, but coerce_to_utf8 was already handling those errors by forcing
the encoding and checking if it was valid, and if it was not it used
replacement characters to supress the errors.

When serializing a command to json, it now calls coerce_to_utf8
repeatedly, so this commit removes all warning/debug logging from the
char_encoding class, and moves it to command processing to avoid
spamming the logs if a single command has multiple lossy string
conversions.

Remove the checksum from commands. This is never checked and was
requiring use to serialize the command the json a second time to create.
After this commit, the "Format payload" profile will actually be
profiling how long it takes the format the command.