(PDB-5554) Wrap pg14 drop of partition tables in db-lock-status
...and try with timeout. Makes the separated drop-partition-tables
function match the locking and timeout behavior of the original
prune-daily-partitions.

It's possible that the lock status used by pe-puppetdb-extensions sync
https://github.com/puppetlabs/pe-puppetdb-extensions/blob/c53a6c614d31a584aba1b67fa26da8ee0c92221c/src/puppetlabs/pe_puppetdb_extensions/sync/services.clj
are no longer needed with the current query-bulldozer? But for now this
behavior is mostly equivalent to what we had prior to the addition of
detaching partitions for pg14 workflows.