(PDB-5554,PDB-5555) Remove access-exclusive locks for detached prune
When pruning reports for pg-14 after having detached the tables, it
should no longer be necessary to acquire locks on the
reports/resource-event tables since they won't be involved when we drop
the detached partitions.

Adding explicit locks for the detached partitions should also be unnecessary
since nothing should be trying to reach them directly.

Acquiring the lock on certnames ensures we get it quickly and is still a
good ordering practice for avoiding deadlocks since generally we acquire
certnames first in other operations.