(maint) Install puppetdb module 7.10.0 on platform5
For 6.x, the acceptance suite running with INSTALL_TYPE upgrade_oldest will
install puppet5 on el7 hosts in order to test upgrading from puppet5.
However it currently installs the puppetlabs-puppetdb module latest to
install puppetdb, and head of that module depends on
puppetlabs-postgresql module <9 which pulls in 8.2.1 which makes use of
Deferred, which isn't available in puppet5...

So if we're installing the puppetdb module for platform5, install
version 7.5.0 which depends on postgresql module <8 that doesn't use
Deferred.