(PDB-5553) Remove foreign keys from report partitions
We have not released this migration, otherwise a separate migration
would be needed.

Postgres foreign keys are implemented with bidirectional triggers this
means that even after we have detached the partition from the parent
table, dropping it will still require access exclusive locks on all the
tables that we reference with foreign keys.