(PDB-5553) Ensure reports are removed when certname purged
Adds a do-hsql function because honeysql outputs standard jdbc format
commands, but only our jdbc query function follows conventions. In
do-prepared, we use non-standard syntax to support SQL as list of
strings, which would misinterpret the output of honeysql/format.

Switch to using honeysql so that we do not have to format our own
in-array queries.

Unify the certname deletion strategy. We currently had two different
strategies for deleting certname data. In delete-certnames, we do a
bottom up deletion, but in the periodic purge method, we rely on deletes
to cascade. This implements a multi-certname delete function that
operates in a more bottom-up fasion (although it is not yet completely
bottom-up there still will be some cascading).

In doing this, purging a number of nodes should be more efficient
because we eliminate cascading to the reports and resource events
partitions, which would be particularly expensive. Future optimizations
would likely benefit from looking at our cascades to catalogs and
factsets.