(PDB-3171) Improve performance for queries with large arrays
Previously, the tree visitor code would *always* replace every node it
visited, regardless of whether the value was actually changed. The code
is intended to allow the visitor to return nil to indicate the node
shouldn't be replaced. In practice, that doesn't actually work for a
couple of reasons. The first is that we weren't diligent about ensuring
that every visitor function returns `nil` when there is no change to be
made. The other is that the `node-visit` function had a bug which would
cause it to always return the original node value if the visitor
function returned `nil`.

The problem is compounded by the fact that whenever a child node is
replaced, the *entire* parent is rewritten. So for large arrays, that
means that we recreate the whole array node once for every child. This
caused significant performance problems for queries with large arrays
(like a large "in" clause), because the parent array is entirely
replaced once for every single one of its children.

We now separately track the "current" node value and the "new" node
value, so that a sequence of visitors that each return `nil` will all be
applied to the initial node value, but we will still return `nil` for
the final replacement value.

Using the following query as a test, this change reduces query time from
~40sec to ~160ms (a 250x improvement).

["from", "nodes",
  ["in", "certname", ["array", [...10 thousand elements...]]]]