(maint) pin hikari to 2.7.4
A recent change in clj-parent https://github.com/puppetlabs/clj-parent/commit/14dd505bd3f3b45d474f9db407a01ba60e86caab
brought in a new version of HikariCP which is built against a newer jdk.

This change means that puppetdb on the 7.x branch is no longer compatible
with jdk8 and would need to be packaged with a newer version of the
jdk.  In order to maintain backwards compatibility with jdk8 packaging,
this pins hikari to 2.7.4 on the 7.x branch. It is not intended to be
on main.