(PDB-5591) Add factset generation to pdb cli generate
Provides several additional parameters to generate factsets alongside
catalogs based on number of facts, and depth.

Adds a baseline facts file to resources to ensure we get a some common
fact paths for hosts.

Mutates some of the baseline, values, then adds or removes facts as
neede to reach requested number of facts.