(PDB-4937) Enforce query timeouts on the db side too
Enforce the query timeouts on the database side by setting a suitable
statement_timeout and idle_in_transaction_timeout in the query's
transaction, and update that timeout between any count query and the
primary query.

Note that postgrsql appears to enforce these timeouts on a
per-"batch" (per execute[1] message) basis, not with respect to the
whole select statement, so at the moment, this is just best-effort,
i.e. with a timeout of one minute, a select could still run for much
longer if each execute (likely of the jdbc fetch-size), finishes in
less than a minute.

Of course if puppetdb itself is functioning correctly, presumably the
timeouts in that side should prevent exceeding the timeout by more
than the time it takes to get a final fetch-size batch.

[1] https://www.postgresql.org/docs/11/protocol-flow.html#PROTOCOL-FLOW-EXT-QUERY