(maint) acceptance tests: simplify and reduce duplication
Remove initialize_repo_on_host, a large function in helper.rb that
performed curl/dpkg/rpm commands to setup the puppet platform repo based
on the platform under test, the puppet repo version, and the nightly
configuration option. All this is also done in the beaker-puppet
library, so use that instead.

Much of the "logic" we still manage now lives in the puppet_repo_version
function. For upgrade_oldest it would still be better to ensure we
installed the oldest agent/server versions as well.