(PDB-5591) Fatten facts to total fact size and summarize
Adds --total-fact-size parameter for increasing the bytes of a factset
to approximately the requested kB. First mutates the baseline facts,
then creates additional facts if requested count is higher, or deletes
down to requested fact count before increasing size of string values to
reach requested fact weight.

Updates summarize to provide a summary of the factsets as well as
catalogs.