(PDB-5591) Guard safe-sample-normal bounds
Calling the safe-sample-normal function with lower bound greater than
the mean or upper bound less than the mean doesn't generally make much
sense. Throw an exception if we hit these cases.

Verify that vary-param works correctly for 0, and that we get an
exception if we try to vary a negative param.