(PDB-5643) Use recursive PQL rules for and/or
Change PQL expr-or and expr-and grammar rules to use a recursive
definition.

By using instaparse's `{ }` operator (meaning zero or more) in
self-referential rules the grammar is ambiguous.

a = 1 or b = 2 or c = 3 can be parsed as something that looks like
[:or
  a = 1
  b = 2
  c = 3]

Or it could be parsed as

[:or
  a = 1
  [:or
    b = 2
    c = 3]]

Both are valid given the self-referential grammar rule, but as clauses
compound the number of valid parsings grows exponentially significantly
affecting performance. As few as 16 or clauses can cause an OOM with 1GB
heap.

The recursive grammar definition means only one (the second example
above) version is valid. Every and/or operator will have two children.