(PE-36120) Do not kill connections when no migrations are pending
Previous behavior would cause compilers to be disconnected from
postgresql ever time the primary's pe-puppetdb was restarted. This was because the
restart invokes the migration code path, which, in PE with a
migrator-account set leads to connection blocking even when the
migration itself is ultimately a noop.

Now it tests whether any migrations are pending before calling
call-with-connections-blocked-during-migration. This change also means
that update-schema will return nil in this case, and initialize-schema
will return false, which is a change from previous noop migrations where
initialize-schema would have returned true.