(PDB-5645) Add test-specific configuration via a ::conf/test map
Make it possible to provide a test-oriented map in the server config
that will be broadly available at run time.

If a [:puppetdb :puppetlabs.puppetdb.config/test] value exists in the
incoming config, pass it straight through to the final configuration,
and then transfer the value into the puppetdb service shared globals
as :puppetlabs.puppetdb.config/test.

This is intended to provide a better alternative to a number of cases
that might otherwise be handled via redefs.