(PDB-5643) Transform back to flattened OR AST expressions
The patch to the pql-grammar.ebnf took care of the ambiguity causing
exponential parsing of deep OR clauses, but produces a nested AST form
that differs from the flat [or 1 2 3 4] structure that was previously
being handed off to the query engine.

Nick Lewis had suggested an additional patch to transform.clj which
would restructure the AST back to the form that the PQL parser was
previously producing so that the rest of the system isn't blind-sided by
a differing query structure. Since the issue we were fixing was in the
initial parse of the PQL, taken together, these patches should fix the
slow parse time without introducing any other novel behavior in the
query engine.