(maint) acceptance: install tzdata-java on el8
With the release of openjdk-headless 11.0.20.0.8-2.el8 tzdata-java
is no longer included as a dependency of openjdk. This is a bug in
RedHat's packages
https://bugzilla.redhat.com/show_bug.cgi?id=2224427

PuppetServer fails to start when the package is missing. EZBake has been
adjusted to depend on the package in future releases, but there's
nothing we can do for older releases, so to test upgrade_latest we must
install the package ourselves.