(maint) Fix cli.benchmark --rand-perc catalog keys
A bug in how catalogs were being mutated was causing benchmark to fail
with --rand-perc sent. The issue was that the puppetdb.catalog.util
mutator functions walk/keywordize-keys the catalog, but other mutations
in benchmark expect string keys. This was leading over --runinterval
time, to catalogs with some duplicate keys (both a "producer_timestamp"
and a :producer_timestamp, for example). And this was eventually causing
errors like:

Exception while submitting command: java.lang.IllegalArgumentException: No implementation of method: :to-date-time of protocol: #'clj-time.coerce/ICoerce found for class: clojure.lang.PersistentArrayMap

That was being thrown when puppetdb.client was hitting a serialized
:producer_timestamp value.

The fix here is just to ensure that we walk/stringify-keys the catalog
after its gone through puppetdb.catalog.util mutation functions.